+++ CHANGELOG.md
@@ -7,9 +7,16 @@
 	- Full support for CubeIDE imports.
 	- Add option to select installation location for the tooling.
 
-## [3.2.7] - 2024-01-09
+## [3.2.7] - 2024-04-28
 ### Added
 - Issue #84 Add -fno-rtti and -fno-exceptions as default for C++ builds.
+- Added an .stm32env file which set-up the paths to openocd and the ARM toolchain. This can also be used to change other things in the build in a local setup.
+- Issue #144 Added a way to change the build folder using the .stm32env file.
+
+### Fix
+- Issue #181 added a way to extract the asmm part of files from the ST Makefile. 
+- Issue #179 now create a prompt to save cortex debug info to the global space, so is should not create a local workspace setting each time.
+- Issue #176 Added a project name field for the blank setup wizard so the project has a default name.
 
 ## [3.2.6] - 2024-01-09
 ### Fixed